fix(pipeline): serializar monto_postulado como float en response schema
Decimal serializa a string en Pydantic v2, rompiendo operaciones numéricas
en el frontend. El request schema conserva Decimal para validar precisión.

+++ backend/app/schemas/pipeline.py
@@ -48,7 +48,7 @@ class PipelineItemListItem(BaseModel):
     score: int | None
     score_justificacion: dict[str, Any] | None
     razon_descarte: str | None
-    monto_postulado: Decimal | None
+    monto_postulado: float | None
     resultado_observaciones: str | None
     detected_by_radar_id: uuid.UUID | None
     notas_count: int
